chore: fix asdf detection

+++ plugins/asdf/asdf.sh
@@ -4,16 +4,17 @@
 ASDF_DIR="${ASDF_DIR:-$HOME/.asdf}"
 ASDF_COMPLETIONS="$ASDF_DIR/completions"
 
-if [[ ! -f "${ASDF_DIR}/asdf.sh" ]] || [[ -f $(which brew) ]]; then
+# If not found, check for Homebrew package
+if [[ ! -f "$ASDF_DIR/asdf.sh" || ! -f "$ASDF_COMPLETIONS/asdf.bash" ]] && [[ -f $(which brew 2>&1) ]]; then
+   ASDF_DIR="$(brew --prefix asdf)"
+   ASDF_COMPLETIONS="$ASDF_DIR/etc/bash_completion.d"
+fi
+
+if ! [[ -f "${ASDF_DIR}/asdf.sh" ]] && [[ -f $(which brew 2>&1) ]]; then
     echo "[asdf] Installing asdf via brew";
     brew install asdf;
 fi
 
-# If not found, check for Homebrew package
-if [[ ! -f "$ASDF_DIR/asdf.sh" || ! -f "$ASDF_COMPLETIONS/asdf.bash" ]] && [[ -f $(which brew 2&>1) ]]; then
-   ASDF_DIR="$(brew --prefix asdf)"
-   ASDF_COMPLETIONS="$ASDF_DIR/etc/bash_completion.d"
-fi
 
 # Load command
 if [[ -f "$ASDF_DIR/asdf.sh" ]]; then
